login and signup work, files renames
