+++ package.json
@@ -21,7 +21,6 @@
     "nodemon": "^1.18.7"
   },
   "dependencies": {
-    
     "config": "^3.2.2",
     "dotenv": "^8.0.0",
     "express": "^4.16.3",
@@ -35,6 +34,5 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/healthy-people/healthy-back-end"
-
   }
 }
